requirements.txt
flask
torch
torchvision
transformers
pillow
together
python-dotenv
requests
